Tests no duplication in loading benchmark times
Test for benchmark time loading is updated. Now the test ensures that
the benchmark times loaded is identical (one-to-one checked) to the raw
yaml file.

Before, the test only guarantees that the benchmark time loaded is a
*subset* of the benchmark times in the raw yaml file. It is possible,
however unplausible, that the Config object loads only a few of the the times
listed in the yaml file, or, worse still, a repetition of benchmark
times. The test now correctly tests this possibility by ensuring that
the data loaded and the raw data are one-to-one equivalent.